requirements.txt
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.2.0
PyQt5==5.15.10
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
pyzbar==0.1.9
